Rework - reorganize; move all experiments and analyses to own library.driven_freq_tuning - add util; define StarkCoefficients dataclasses and util functions - separation; delegated the role of coefficient manipulation from the analysis class to util functions
