+++ docs/apidocs/mod_driven_freq_tuning.rst
@@ -1,0 +1,6 @@
+.. _qiskit-experiments-driven-freq-tuning:
+
+.. automodule:: qiskit_experiments.library.driven_freq_tuning
+   :no-members:
+   :no-inherited-members:
+   :no-special-members:
