+++ qiskit_experiments/library/__init__.py
@@ -76,8 +76,9 @@ Some experiments also have a calibration experiment version.
     ~characterization.FineXDrag
     ~characterization.FineSXDrag
     ~characterization.MultiStateDiscrimination
-    ~characterization.StarkRamseyXY
-    ~characterization.StarkRamseyXYAmpScan
+    ~driven_freq_tuning.StarkRamseyXY
+    ~driven_freq_tuning.StarkRamseyXYAmpScan
+    ~driven_freq_tuning.StarkP1Spectroscopy
 
 .. _characterization two qubits:
 
@@ -160,7 +161,6 @@ from .calibration import (
 )
 from .characterization import (
     T1,
-    StarkP1Spectroscopy,
     T2Hahn,
     T2Ramsey,
     Tphi,
@@ -187,8 +187,6 @@ from .characterization import (
     CorrelatedReadoutError,
     ZZRamsey,
     MultiStateDiscrimination,
-    StarkRamseyXY,
-    StarkRamseyXYAmpScan,
 )
 from .randomized_benchmarking import StandardRB, InterleavedRB
 from .tomography import (
@@ -199,6 +197,11 @@ from .tomography import (
     MitigatedProcessTomography,
 )
 from .quantum_volume import QuantumVolume
+from .driven_freq_tuning import (
+    StarkRamseyXY,
+    StarkRamseyXYAmpScan,
+    StarkP1Spectroscopy,
+)
 
 # Experiment Sub-modules
 from . import calibration
