+++ test/library/driven_freq_tuning/test_stark_p1_spect.py
@@ -23,7 +23,8 @@ from qiskit.providers.fake_provider import FakeHanoiV2
 
 from qiskit_experiments.framework import ExperimentData, AnalysisResultData
 from qiskit_experiments.library import StarkP1Spectroscopy
-from qiskit_experiments.library.characterization.analysis import StarkP1SpectAnalysis
+from qiskit_experiments.library.driven_freq_tuning.analyses import StarkP1SpectAnalysis
+from qiskit_experiments.library.driven_freq_tuning import coefficient_utils as util
 from qiskit_experiments.test import FakeService
 
 
@@ -47,49 +48,6 @@ class StarkP1SpectAnalysisReturnXvals(StarkP1SpectAnalysis):
 @ddt
 class TestStarkP1Spectroscopy(QiskitExperimentsTestCase):
     """Test case for the Stark P1 Spectroscopy experiment."""
-
-    @staticmethod
-    def create_service_helper(
-        pos_coef_o1: float,
-        pos_coef_o2: float,
-        pos_coef_o3: float,
-        neg_coef_o1: float,
-        neg_coef_o2: float,
-        neg_coef_o3: float,
-        ferr: float,
-        qubit: int,
-        backend_name: str,
-    ):
-        """A helper method to create service with analysis results."""
-        service = FakeService()
-
-        service.create_experiment(
-            experiment_type="StarkRamseyXYAmpScan",
-            backend_name=backend_name,
-            experiment_id="123456789",
-        )
-
-        coeffs = {
-            "stark_pos_coef_o1": pos_coef_o1,
-            "stark_pos_coef_o2": pos_coef_o2,
-            "stark_pos_coef_o3": pos_coef_o3,
-            "stark_neg_coef_o1": neg_coef_o1,
-            "stark_neg_coef_o2": neg_coef_o2,
-            "stark_neg_coef_o3": neg_coef_o3,
-            "stark_ferr": ferr,
-        }
-        for i, (key, value) in enumerate(coeffs.items()):
-            service.create_analysis_result(
-                experiment_id="123456789",
-                result_data={"value": value},
-                result_type=key,
-                device_components=[f"Q{qubit}"],
-                tags=[],
-                quality="Good",
-                verified=False,
-                result_id=str(i),
-            )
-        return service
 
     def test_linear_spaced_parameters(self):
         """Test generating parameters with linear spacing."""
@@ -126,96 +84,59 @@ class TestStarkP1Spectroscopy(QiskitExperimentsTestCase):
             exp.set_experiment_options(spacing="invalid_option")
 
     def test_raises_scanning_frequency_without_service(self):
-        """Test raises error when frequency is set without having service or coefficients set."""
+        """Test raises error when frequency is set without no coefficients.
+
+        This covers following situations:
+        - stark_coefficients options is None
+        - backend object doesn't provide experiment service
+        """
         exp = StarkP1Spectroscopy((0,), backend=FakeHanoiV2())
         exp.set_experiment_options(
             xvals=[-100e6, -50e6, 0, 50e6, 100e6],
             xval_type="frequency",
-            stark_coefficients="latest",
         )
         with self.assertRaises(RuntimeError):
             exp.parameters()
 
-    @named_data(
-        ["ordinary", 5e6, 200e6, -50e6, 5e6, -180e6, -40e6, 100e3],
-        ["asymmetric_inflection_1st_ord", 10e6, 200e6, -20e6, -50e6, -180e6, -20e6, -10e6],
-        ["inflection_3st_ord", 10e6, 200e6, -80e6, 80e6, -180e6, -200e6, 100e3],
-    )
-    @unpack
-    def test_scanning_frequency(self, po1, po2, po3, no1, no2, no3, ferr):
-        """Test scanning frequency with experiment service.
-
-        This is a sort of round-trip test.
-        We generate amplitudes from frequencies through the experiment class.
-        These amplitudes are converted into frequencies again with the same coefficients.
-        The two sets of frequencies must be consistent.
-        """
-        service = self.create_service_helper(po1, po2, po3, no1, no2, no3, ferr, 0, "fake_hanoi")
-
-        exp = StarkP1Spectroscopy((0,), backend=FakeHanoiV2())
-
-        ref_freqs = np.linspace(-70e6, 70e6, 31)
-        exp.set_experiment_options(
-            xvals=ref_freqs,
-            xval_type="frequency",
-            service=service,
-        )
-
-        amplitudes = exp.parameters()
-
-        # Compute frequency from parameter values with the same coefficient
-        analysis = StarkP1SpectAnalysis()
-        coeffs = analysis.retrieve_coefficients_from_service(service, 0, "fake_hanoi")
-        frequencies = analysis._convert_axis(amplitudes, coeffs)
-
-        np.testing.assert_array_almost_equal(frequencies, ref_freqs)
-
     def test_scanning_frequency_with_coeffs(self):
-        """Test scanning frequency with manually provided Stark coefficients.
-
-        This is just a difference of API from the test_scanning_frequency.
-        Data driven test is omitted here.
-        """
-        po1, po2, po3, no1, no2, no3, ferr = 5e6, 200e6, -50e6, 5e6, -180e6, -40e6, 100e3
+        """Test scanning frequency with manually provided Stark coefficients."""
+        coeffs = util.StarkCoefficients(
+            pos_coef_o1=5e6,
+            pos_coef_o2=200e6,
+            pos_coef_o3=-50e6,
+            neg_coef_o1=5e6,
+            neg_coef_o2=-180e6,
+            neg_coef_o3=-40e6,
+            offset=100e3,
+        )
         exp = StarkP1Spectroscopy((0,), backend=FakeHanoiV2())
 
         ref_amps = np.array([-0.50, -0.25, 0.0, 0.25, 0.50], dtype=float)
-        test_freqs = np.where(
-            ref_amps > 0,
-            po1 * ref_amps + po2 * ref_amps**2 + po3 * ref_amps**3 + ferr,
-            no1 * ref_amps + no2 * ref_amps**2 + no3 * ref_amps**3 + ferr,
-        )
+        test_freqs = util.convert_amp_to_freq(ref_amps, coeffs)
         exp.set_experiment_options(
             xvals=test_freqs,
             xval_type="frequency",
-            stark_coefficients={
-                "stark_pos_coef_o1": po1,
-                "stark_pos_coef_o2": po2,
-                "stark_pos_coef_o3": po3,
-                "stark_neg_coef_o1": no1,
-                "stark_neg_coef_o2": no2,
-                "stark_neg_coef_o3": no3,
-                "stark_ferr": ferr,
-            },
+            stark_coefficients=coeffs,
         )
         params = exp.parameters()
         np.testing.assert_array_almost_equal(params, ref_amps)
 
     def test_scaning_frequency_around_zero(self):
         """Test scanning frequency around zero."""
+        coeffs = util.StarkCoefficients(
+            pos_coef_o1=5e6,
+            pos_coef_o2=100e6,
+            pos_coef_o3=10e6,
+            neg_coef_o1=-5e6,
+            neg_coef_o2=-100e6,
+            neg_coef_o3=-10e6,
+            offset=500e3,
+        )
         exp = StarkP1Spectroscopy((0,), backend=FakeHanoiV2())
         exp.set_experiment_options(
             xvals=[0, 500e3],
             xval_type="frequency",
-            stark_coefficients={
-                "stark_pos_coef_o1": 5e6,
-                "stark_pos_coef_o2": 100e6,
-                "stark_pos_coef_o3": 10e6,
-                "stark_neg_coef_o1": -5e6,
-                "stark_neg_coef_o2": -100e6,
-                "stark_neg_coef_o3": -10e6,
-                "stark_ferr": 500e3,
-            },
+            stark_coefficients=coeffs,
         )
         params = exp.parameters()
         # Frequency offset is 500 kHz and we need negative shift to tune frequency at zero.
@@ -278,50 +199,6 @@ class TestStarkP1Spectroscopy(QiskitExperimentsTestCase):
         self.assertEqual(circs[0], qc1)
         self.assertEqual(circs[1], qc2)
 
-    def test_retrieve_coefficients(self):
-        """Test retrieving Stark coefficients from the experiment service."""
-        po1, po2, po3, no1, no2, no3, ferr = 5e6, 200e6, -50e6, 5e6, -180e6, -40e6, 100e3
-        service = self.create_service_helper(po1, po2, po3, no1, no2, no3, ferr, 0, "fake_hanoi")
-
-        retrieved_coeffs = StarkP1SpectAnalysis.retrieve_coefficients_from_service(
-            service=service,
-            qubit=0,
-            backend="fake_hanoi",
-        )
-        self.assertDictEqual(
-            retrieved_coeffs,
-            {
-                "stark_pos_coef_o1": po1,
-                "stark_pos_coef_o2": po2,
-                "stark_pos_coef_o3": po3,
-                "stark_neg_coef_o1": no1,
-                "stark_neg_coef_o2": no2,
-                "stark_neg_coef_o3": no3,
-                "stark_ferr": ferr,
-            },
-        )
-
-    @named_data(
-        ["ordinary", 5e6, 200e6, -50e6, 5e6, -180e6, -40e6, 100e3],
-        ["asymmetric_inflection_1st_ord", 10e6, 200e6, -20e6, -50e6, -180e6, -20e6, -10e6],
-        ["inflection_3st_ord", 10e6, 200e6, -80e6, 80e6, -180e6, -200e6, 100e3],
-    )
-    @unpack
-    def test_estimating_minmax_frequency(self, po1, po2, po3, no1, no2, no3, ferr):
-        """Test computing the minimum and maximum frequency within the amplitude budget."""
-        service = self.create_service_helper(po1, po2, po3, no1, no2, no3, ferr, 0, "fake_hanoi")
-        analysis = StarkP1SpectAnalysis()
-
-        coeffs = analysis.retrieve_coefficients_from_service(service, 0, "fake_hanoi")
-        # pylint: disable=unbalanced-tuple-unpacking
-        minf, maxf = analysis.estimate_minmax_frequencies(coeffs, (-0.9, 0.9))
-
-        amps = np.linspace(-0.9, 0.9, 101)
-        freqs = analysis._convert_axis(amps, coeffs)
-
-        self.assertAlmostEqual(minf, min(freqs), delta=1e6)
-        self.assertAlmostEqual(maxf, max(freqs), delta=1e6)
-
     def test_running_analysis_without_service(self):
         """Test running analysis without setting service to the experiment data.
 
@@ -349,15 +226,42 @@ class TestStarkP1Spectroscopy(QiskitExperimentsTestCase):
 
         This must convert x-axis into frequencies with the Stark coefficients.
         """
-        service = self.create_service_helper(po1, po2, po3, no1, no2, no3, ferr, 0, "fake_hanoi")
+        mock_experiment_id = "6453f3d1-04ef-4e3b-82c6-1a92e3e066eb"
+        mock_result_id = "d067ae34-96db-4e8e-adc8-030305d3d404"
+        mock_backend = FakeHanoiV2().name
+
+        coeffs = util.StarkCoefficients(
+            pos_coef_o1=po1,
+            pos_coef_o2=po2,
+            pos_coef_o3=po3,
+            neg_coef_o1=no1,
+            neg_coef_o2=no2,
+            neg_coef_o3=no3,
+            offset=ferr,
+        )
+
+        service = FakeService()
+        service.create_experiment(
+            experiment_type="StarkRamseyXYAmpScan",
+            backend_name=mock_backend,
+            experiment_id=mock_experiment_id,
+        )
+        service.create_analysis_result(
+            experiment_id=mock_experiment_id,
+            result_data={"value": coeffs},
+            result_type="stark_coefficients",
+            device_components=["Q0"],
+            tags=[],
+            quality="Good",
+            verified=False,
+            result_id=mock_result_id,
+        )
+
         analysis = StarkP1SpectAnalysisReturnXvals()
 
         xvals = np.linspace(-1, 1, 11)
-        ref_xvals = np.where(
-            xvals > 0,
-            po1 * xvals + po2 * xvals**2 + po3 * xvals**3 + ferr,
-            no1 * xvals + no2 * xvals**2 + no3 * xvals**3 + ferr,
-        )
+        ref_fvals = util.convert_amp_to_freq(xvals, coeffs)
+
         exp_data = ExperimentData(
             service=service,
             backend=FakeHanoiV2(),
@@ -365,9 +269,9 @@ class TestStarkP1Spectroscopy(QiskitExperimentsTestCase):
         exp_data.metadata.update({"physical_qubits": [0]})
         for x in xvals:
             exp_data.add_data({"counts": {"0": 1000, "1": 0}, "metadata": {"xval": x}})
-        analysis.run(exp_data, replace_results=True)
-        test_xvals = exp_data.analysis_results("xvals").value
-        np.testing.assert_array_almost_equal(test_xvals, ref_xvals)
+        analysis.run(exp_data, replace_results=True).block_for_results()
+        test_fvals = exp_data.analysis_results("xvals").value
+        np.testing.assert_array_almost_equal(test_fvals, ref_fvals)
 
     def test_running_analysis_with_user_provided_coeffs(self):
         """Test running analysis by manually providing Stark coefficients.
@@ -376,30 +280,25 @@ class TestStarkP1Spectroscopy(QiskitExperimentsTestCase):
         This is just a difference of API from the test_running_analysis_with_service.
         Data driven test is omitted here.
         """
-        po1, po2, po3, no1, no2, no3, ferr = 5e6, 200e6, -50e6, 5e6, -180e6, -40e6, 100e3
+        coeffs = util.StarkCoefficients(
+            pos_coef_o1=5e6,
+            pos_coef_o2=200e6,
+            pos_coef_o3=-50e6,
+            neg_coef_o1=5e6,
+            neg_coef_o2=-180e6,
+            neg_coef_o3=-40e6,
+            offset=100e3,
+        )
 
         analysis = StarkP1SpectAnalysisReturnXvals()
-        analysis.set_options(
-            stark_coefficients={
-                "stark_pos_coef_o1": po1,
-                "stark_pos_coef_o2": po2,
-                "stark_pos_coef_o3": po3,
-                "stark_neg_coef_o1": no1,
-                "stark_neg_coef_o2": no2,
-                "stark_neg_coef_o3": no3,
-                "stark_ferr": ferr,
-            }
-        )
+        analysis.set_options(stark_coefficients=coeffs)
 
         xvals = np.linspace(-1, 1, 11)
-        ref_xvals = np.where(
-            xvals > 0,
-            po1 * xvals + po2 * xvals**2 + po3 * xvals**3 + ferr,
-            no1 * xvals + no2 * xvals**2 + no3 * xvals**3 + ferr,
-        )
+        ref_fvals = util.convert_amp_to_freq(xvals, coeffs)
+
         exp_data = ExperimentData()
         for x in xvals:
             exp_data.add_data({"counts": {"0": 1000, "1": 0}, "metadata": {"xval": x}})
-        analysis.run(exp_data, replace_results=True)
-        test_xvals = exp_data.analysis_results("xvals").value
-        np.testing.assert_array_almost_equal(test_xvals, ref_xvals)
+        analysis.run(exp_data, replace_results=True).block_for_results()
+        test_fvals = exp_data.analysis_results("xvals").value
+        np.testing.assert_array_almost_equal(test_fvals, ref_fvals)
